requirements.txt
pyzmq==17.0.0
redis==2.10.6
